extract OnDiskVectors to top level class so we can use it in debugging tools
